Milk 클래스 fatContent Double로 수정

+++ VendingMachine/VendingMachine/Class/Milk/ChocolateMilk.swift
@@ -11,7 +11,7 @@ import Foundation
 class ChocolateMilk: Milk {
     private let chocolateContent: Int
     
-    init(brand: String, capacity: Int, price: Int, name: String, date: Date, fatContent: Int, chocolateContent: Int) {
+    init(brand: String, capacity: Int, price: Int, name: String, date: Date, fatContent: Double, chocolateContent: Int) {
         self.chocolateContent = chocolateContent
         
         super.init(brand: brand, capacity: capacity, price: price, name: name, date: date, fatContent: fatContent)
