+++ VendingMachine/VendingMachine/Class/Milk/StrawberryMilk.swift
@@ -11,7 +11,7 @@ import Foundation
 class StrawberryMilk: Milk {
     private let strawberryContent: Int
     
-    init(brand: String, capacity: Int, price: Int, name: String, date: Date, fatContent: Int, strawberryContent: Int) {
+    init(brand: String, capacity: Int, price: Int, name: String, date: Date, fatContent: Double, strawberryContent: Int) {
         self.strawberryContent = strawberryContent
         
         super.init(brand: brand, capacity: capacity, price: price, name: name, date: date, fatContent: fatContent)
